Add the initial version of SblPod Tool
This tool could work as a standalone tool to generate PodImage.
To generate SubRegion, it has a builtin region list to build
subRegion for test. In the next, need provide by BoardConfig.py
or separate file to provide region list.
This tool depends on python, openssl, lzmacompress and lz4compress.

Signed-off-by: Guo Dong <guo.dong@intel.com>